refactor(library): extract remaining screen components and share badges, search field and list animation
