+++ app/src/main/java/io/github/aedev/flow/ui/components/library/DownloadRows.kt
@@ -1,11 +1,15 @@
 package io.github.aedev.flow.ui.components.library
 
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
@@ -14,12 +18,17 @@ import androidx.compose.material.icons.filled.PlayArrow
 import androidx.compose.material.icons.outlined.Delete
 import androidx.compose.material.icons.outlined.MusicNote
 import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.FilledTonalButton
 import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
+import androidx.compose.material3.pulltorefresh.PullToRefreshBox
+import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -32,6 +41,11 @@ import io.github.aedev.flow.data.local.entity.DownloadItemStatus
 import io.github.aedev.flow.data.local.entity.DownloadWithItems
 import io.github.aedev.flow.data.music.DownloadedTrack
 import io.github.aedev.flow.data.video.DownloadedVideo
+import io.github.aedev.flow.ui.components.library.ActiveDownloadRow
+import io.github.aedev.flow.ui.components.library.DownloadsEmptyState
+import io.github.aedev.flow.ui.components.library.DownloadsSectionHeader
+import io.github.aedev.flow.ui.components.library.MusicDownloadRow
+import io.github.aedev.flow.ui.components.library.VideoDownloadRow
 import io.github.aedev.flow.ui.components.shared.ArtworkThumbnail
 import io.github.aedev.flow.ui.components.shared.ExplicitBadge
 import io.github.aedev.flow.ui.components.shared.FlowEmptyState
@@ -39,6 +53,7 @@ import io.github.aedev.flow.ui.components.shared.MediaKind
 import io.github.aedev.flow.ui.components.shared.MediaRow
 import io.github.aedev.flow.ui.components.shared.MediaRowAction
 import io.github.aedev.flow.ui.components.shared.MediaThumbnail
+import io.github.aedev.flow.ui.components.shared.animateMediaListItem
 
 private val ProgressBarHeight: Dp = 3.dp
 private val SectionHeaderPadding = PaddingValues(horizontal = 16.dp, vertical = 6.dp)
@@ -241,5 +256,134 @@ internal fun DownloadsEmptyState(
                 }
             },
         )
+    }
+}
+
+private val ListContentPadding = PaddingValues(vertical = 8.dp)
+private val ListItemSpacing = 2.dp
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun VideosDownloadsList(
+    videos: List<DownloadedVideo>,
+    incompleteDownloads: List<DownloadWithItems>,
+    progressMap: Map<String, Float>,
+    mergingVideoIds: Set<String>,
+    isRefreshing: Boolean,
+    onRefresh: () -> Unit,
+    onVideoClick: (List<DownloadedVideo>, Int) -> Unit,
+    onDeleteClick: (String, String) -> Unit,
+    onPauseClick: (String) -> Unit,
+    onResumeClick: (String) -> Unit,
+    onHomeClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    val pullState = rememberPullToRefreshState()
+    PullToRefreshBox(
+        isRefreshing = isRefreshing,
+        onRefresh = onRefresh,
+        state = pullState,
+        modifier = modifier.fillMaxSize(),
+    ) {
+        if (videos.isEmpty() && incompleteDownloads.isEmpty()) {
+            DownloadsEmptyState(kind = MediaKind.Videos, onHomeClick = onHomeClick)
+        } else {
+            LazyColumn(
+                modifier = Modifier.fillMaxSize(),
+                contentPadding = ListContentPadding,
+                verticalArrangement = Arrangement.spacedBy(ListItemSpacing),
+            ) {
+                if (incompleteDownloads.isNotEmpty()) {
+                    item(key = "section_active", contentType = "section") {
+                        DownloadsSectionHeader(
+                            text = stringResource(R.string.section_incomplete_downloads),
+                            color = MaterialTheme.colorScheme.primary,
+                        )
+                    }
+                    items(
+                        items = incompleteDownloads,
+                        key = { "active_${it.download.videoId}" },
+                        contentType = { "active" },
+                    ) { download ->
+                        ActiveDownloadRow(
+                            download = download,
+                            progressMap = progressMap,
+                            isMerging = download.download.videoId in mergingVideoIds,
+                            onPauseClick = { onPauseClick(download.download.videoId) },
+                            onResumeClick = { onResumeClick(download.download.videoId) },
+                            onDeleteClick = {
+                                onDeleteClick(download.download.videoId, download.download.title)
+                            },
+                            modifier = animateMediaListItem(),
+                        )
+                    }
+                    if (videos.isNotEmpty()) {
+                        item(key = "section_completed", contentType = "section") {
+                            DownloadsSectionHeader(
+                                text = stringResource(R.string.section_completed),
+                                color = MaterialTheme.colorScheme.onSurfaceVariant,
+                            )
+                        }
+                    }
+                }
+                itemsIndexed(
+                    items = videos,
+                    key = { _, video -> video.video.id },
+                    contentType = { _, _ -> "video" },
+                ) { index, video ->
+                    VideoDownloadRow(
+                        video = video,
+                        onClick = { onVideoClick(videos, index) },
+                        onDeleteClick = { onDeleteClick(video.video.id, video.video.title) },
+                        modifier = animateMediaListItem(),
+                    )
+                }
+            }
+        }
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun MusicDownloadsList(
+    tracks: List<DownloadedTrack>,
+    isRefreshing: Boolean,
+    onRefresh: () -> Unit,
+    onMusicClick: (List<DownloadedTrack>, Int) -> Unit,
+    onDeleteClick: (String, String) -> Unit,
+    onHomeClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    val pullState = rememberPullToRefreshState()
+    PullToRefreshBox(
+        isRefreshing = isRefreshing,
+        onRefresh = onRefresh,
+        state = pullState,
+        modifier = modifier.fillMaxSize(),
+    ) {
+        if (tracks.isEmpty()) {
+            DownloadsEmptyState(kind = MediaKind.Music, onHomeClick = onHomeClick)
+        } else {
+            LazyColumn(
+                modifier = Modifier.fillMaxSize(),
+                contentPadding = ListContentPadding,
+                verticalArrangement = Arrangement.spacedBy(ListItemSpacing),
+            ) {
+                itemsIndexed(
+                    items = tracks,
+                    key = { _, track -> track.track.videoId },
+                    contentType = { _, _ -> "track" },
+                ) { index, downloadedTrack ->
+                    MusicDownloadRow(
+                        downloadedTrack = downloadedTrack,
+                        onClick = { onMusicClick(tracks, index) },
+                        onDeleteClick = {
+                            onDeleteClick(downloadedTrack.track.videoId, downloadedTrack.track.title)
+                        },
+                        modifier = animateMediaListItem(),
+                    )
+                }
+            }
+        }
     }
 }
