+++ app/src/main/java/io/github/aedev/flow/ui/components/music/search/MusicSearchComponents.kt
@@ -8,7 +8,6 @@ package io.github.aedev.flow.ui.components.music.search
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyRow
@@ -16,7 +15,6 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.rounded.ArrowBack
 import androidx.compose.material.icons.rounded.ArrowOutward
-import androidx.compose.material.icons.rounded.Close
 import androidx.compose.material.icons.rounded.Mic
 import androidx.compose.material.icons.rounded.Search
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -25,7 +23,6 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.SearchBarDefaults
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -39,6 +36,7 @@ import androidx.compose.ui.unit.dp
 import io.github.aedev.flow.R
 import io.github.aedev.flow.innertube.YouTube.SearchFilter
 import io.github.aedev.flow.ui.components.shared.FlowFilterChip
+import io.github.aedev.flow.ui.components.shared.FlowSearchField
 import io.github.aedev.flow.ui.theme.Dimensions
 
 /**
@@ -57,19 +55,18 @@ fun MusicSearchBar(
     focusRequester: FocusRequester = remember { FocusRequester() },
 ) {
     Surface(color = MaterialTheme.colorScheme.background) {
-        SearchBarDefaults.InputField(
+        FlowSearchField(
             query = query,
             onQueryChange = onQueryChange,
-            onSearch = { onSearch() },
+            placeholder = stringResource(R.string.search_music_placeholder),
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(horizontal = 16.dp, vertical = 8.dp),
+            onSearch = onSearch,
+            onClear = onClearClick,
             expanded = true,
-            onExpandedChange = {},
-            placeholder = {
-                Text(
-                    text = stringResource(R.string.search_music_placeholder),
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-            },
+            focusRequester = focusRequester,
             leadingIcon = {
                 IconButton(onClick = onBackClick) {
                     Icon(
@@ -78,29 +75,14 @@ fun MusicSearchBar(
                     )
                 }
             },
-            trailingIcon = {
-                Row {
-                    if (query.isNotEmpty()) {
-                        IconButton(onClick = onClearClick) {
-                            Icon(
-                                imageVector = Icons.Rounded.Close,
-                                contentDescription = stringResource(R.string.clear),
-                            )
-                        }
-                    }
-                    IconButton(onClick = onVoiceSearchClick) {
-                        Icon(
-                            imageVector = Icons.Rounded.Mic,
-                            contentDescription = stringResource(R.string.voice_search_cd),
-                        )
-                    }
+            trailingContent = {
+                IconButton(onClick = onVoiceSearchClick) {
+                    Icon(
+                        imageVector = Icons.Rounded.Mic,
+                        contentDescription = stringResource(R.string.voice_search_cd),
+                    )
                 }
             },
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp, vertical = 8.dp)
-                    .focusRequester(focusRequester),
         )
     }
 }
