+++ app/src/main/java/io/github/aedev/flow/ui/screens/library/DownloadsScreen.kt
@@ -6,20 +6,13 @@ import android.os.Build
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.animation.Crossfade
-import androidx.compose.animation.core.EaseInCubic
 import androidx.compose.animation.core.EaseOutCubic
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.spring
 import androidx.compose.animation.core.tween
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Delete
 import androidx.compose.material3.AlertDialog
@@ -30,8 +23,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
-import androidx.compose.material3.pulltorefresh.PullToRefreshBox
-import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -49,20 +40,13 @@ import androidx.core.content.ContextCompat
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import io.github.aedev.flow.R
-import io.github.aedev.flow.data.local.entity.DownloadWithItems
 import io.github.aedev.flow.data.music.DownloadedTrack
 import io.github.aedev.flow.data.video.DownloadedVideo
 import io.github.aedev.flow.ui.components.layout.topbar.FlowTopBar
-import io.github.aedev.flow.ui.components.library.ActiveDownloadRow
-import io.github.aedev.flow.ui.components.library.DownloadsEmptyState
-import io.github.aedev.flow.ui.components.library.DownloadsSectionHeader
-import io.github.aedev.flow.ui.components.library.MusicDownloadRow
-import io.github.aedev.flow.ui.components.library.VideoDownloadRow
+import io.github.aedev.flow.ui.components.library.MusicDownloadsList
+import io.github.aedev.flow.ui.components.library.VideosDownloadsList
 import io.github.aedev.flow.ui.components.shared.MediaKind
 import io.github.aedev.flow.ui.components.shared.MediaKindSelector
-
-private val ListContentPadding = PaddingValues(vertical = 8.dp)
-private val ListItemSpacing = 2.dp
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -259,144 +243,3 @@ private data class PendingDeletion(
     val title: String,
     val kind: MediaKind,
 )
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun VideosDownloadsList(
-    videos: List<DownloadedVideo>,
-    incompleteDownloads: List<DownloadWithItems>,
-    progressMap: Map<String, Float>,
-    mergingVideoIds: Set<String>,
-    isRefreshing: Boolean,
-    onRefresh: () -> Unit,
-    onVideoClick: (List<DownloadedVideo>, Int) -> Unit,
-    onDeleteClick: (String, String) -> Unit,
-    onPauseClick: (String) -> Unit,
-    onResumeClick: (String) -> Unit,
-    onHomeClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val pullState = rememberPullToRefreshState()
-    PullToRefreshBox(
-        isRefreshing = isRefreshing,
-        onRefresh = onRefresh,
-        state = pullState,
-        modifier = modifier.fillMaxSize(),
-    ) {
-        if (videos.isEmpty() && incompleteDownloads.isEmpty()) {
-            DownloadsEmptyState(kind = MediaKind.Videos, onHomeClick = onHomeClick)
-        } else {
-            LazyColumn(
-                modifier = Modifier.fillMaxSize(),
-                contentPadding = ListContentPadding,
-                verticalArrangement = Arrangement.spacedBy(ListItemSpacing),
-            ) {
-                if (incompleteDownloads.isNotEmpty()) {
-                    item(key = "section_active", contentType = "section") {
-                        DownloadsSectionHeader(
-                            text = stringResource(R.string.section_incomplete_downloads),
-                            color = MaterialTheme.colorScheme.primary,
-                        )
-                    }
-                    items(
-                        items = incompleteDownloads,
-                        key = { "active_${it.download.videoId}" },
-                        contentType = { "active" },
-                    ) { download ->
-                        ActiveDownloadRow(
-                            download = download,
-                            progressMap = progressMap,
-                            isMerging = download.download.videoId in mergingVideoIds,
-                            onPauseClick = { onPauseClick(download.download.videoId) },
-                            onResumeClick = { onResumeClick(download.download.videoId) },
-                            onDeleteClick = {
-                                onDeleteClick(download.download.videoId, download.download.title)
-                            },
-                            modifier =
-                                Modifier.animateItem(
-                                    fadeInSpec = tween(300, easing = EaseOutCubic),
-                                    fadeOutSpec = tween(200, easing = EaseInCubic),
-                                    placementSpec = spring(dampingRatio = 0.8f, stiffness = Spring.StiffnessLow),
-                                ),
-                        )
-                    }
-                    if (videos.isNotEmpty()) {
-                        item(key = "section_completed", contentType = "section") {
-                            DownloadsSectionHeader(
-                                text = stringResource(R.string.section_completed),
-                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                            )
-                        }
-                    }
-                }
-                itemsIndexed(
-                    items = videos,
-                    key = { _, video -> video.video.id },
-                    contentType = { _, _ -> "video" },
-                ) { index, video ->
-                    VideoDownloadRow(
-                        video = video,
-                        onClick = { onVideoClick(videos, index) },
-                        onDeleteClick = { onDeleteClick(video.video.id, video.video.title) },
-                        modifier =
-                            Modifier.animateItem(
-                                fadeInSpec = tween(300, easing = EaseOutCubic),
-                                fadeOutSpec = tween(200, easing = EaseInCubic),
-                                placementSpec = spring(dampingRatio = 0.8f, stiffness = Spring.StiffnessLow),
-                            ),
-                    )
-                }
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun MusicDownloadsList(
-    tracks: List<DownloadedTrack>,
-    isRefreshing: Boolean,
-    onRefresh: () -> Unit,
-    onMusicClick: (List<DownloadedTrack>, Int) -> Unit,
-    onDeleteClick: (String, String) -> Unit,
-    onHomeClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val pullState = rememberPullToRefreshState()
-    PullToRefreshBox(
-        isRefreshing = isRefreshing,
-        onRefresh = onRefresh,
-        state = pullState,
-        modifier = modifier.fillMaxSize(),
-    ) {
-        if (tracks.isEmpty()) {
-            DownloadsEmptyState(kind = MediaKind.Music, onHomeClick = onHomeClick)
-        } else {
-            LazyColumn(
-                modifier = Modifier.fillMaxSize(),
-                contentPadding = ListContentPadding,
-                verticalArrangement = Arrangement.spacedBy(ListItemSpacing),
-            ) {
-                itemsIndexed(
-                    items = tracks,
-                    key = { _, track -> track.track.videoId },
-                    contentType = { _, _ -> "track" },
-                ) { index, downloadedTrack ->
-                    MusicDownloadRow(
-                        downloadedTrack = downloadedTrack,
-                        onClick = { onMusicClick(tracks, index) },
-                        onDeleteClick = {
-                            onDeleteClick(downloadedTrack.track.videoId, downloadedTrack.track.title)
-                        },
-                        modifier =
-                            Modifier.animateItem(
-                                fadeInSpec = tween(300, easing = EaseOutCubic),
-                                fadeOutSpec = tween(200, easing = EaseInCubic),
-                                placementSpec = spring(dampingRatio = 0.8f, stiffness = Spring.StiffnessLow),
-                            ),
-                    )
-                }
-            }
-        }
-    }
-}
