+++ app/src/main/java/io/github/aedev/flow/ui/screens/music/PlaylistPage.kt
@@ -398,11 +398,7 @@ fun PlaylistPage(
                                     thumbnailOverlay = {
                                         ThumbnailWatchProgress(
                                             videoId = track.videoId,
-                                            modifier =
-                                                Modifier
-                                                    .align(Alignment.BottomStart)
-                                                    .fillMaxWidth()
-                                                    .height(3.dp),
+                                            modifier = Modifier.align(Alignment.BottomStart),
                                         )
                                     },
                                     trailingContent =
